refactor: extract global-state event loop
