+++ src/global_state.rs
@@ -1,5 +1,6 @@
 mod diagnostics;
 mod dispatcher;
+pub(crate) mod event_loop;
 mod handlers;
 pub mod main_loop;
 mod mem_docs;
